Move SQL error formatting code to utils.
